Fix edge cases where card is smaller than svg
The mailto and tel links being smaller than the social media site links
meant that cards with both the mailto and tel links but missing one or
two social site links would have a card smaller than the generated svg
image. This commit better matches the small links to the rest by
increasing the font size of and adding a margin to these smaller links.