Fix illegible Snapchat text, consolidate styling
Instead of setting the background colours for the social media sections
in two separate places they are now set only in main.css - this makes
any updating of the colours easier. Required setting the 'name' attribute
for each in card.js but allowed the removal of the 'color' array and
setting thereof. These colours were also updated to match their websites
colours and the barely legible white text on yellow background of
Snapchat has been fixed.